Moving code to zipstep class
