+++ src/main/java/org/jenkinsci/plugins/pipeline/utility/steps/AbstractFileCompressStep.java
@@ -8,8 +8,6 @@ public abstract class AbstractFileCompressStep extends AbstractFileStep {
     private String exclude;
     private boolean archive = false;
     private boolean overwrite = false;
-    private boolean defaultExcludes = true;
-
     /**
      * The relative path of the base directory to create the archive from.
      * Leave empty to create from the current working directory.
@@ -113,27 +111,6 @@ public abstract class AbstractFileCompressStep extends AbstractFileStep {
     @DataBoundSetter
     public void setOverwrite(boolean overwrite) {
         this.overwrite = overwrite;
-    }
-
-    /**
-     * <a href="https://ant.apache.org/manual/dirtasks.html#defaultexcludes" target="_blank">Ant style pattern</a>
-     * of files to disable default excludes from the archive.
-     *
-     * @return the defaultExcludes boolean
-     */
-    public boolean isDefaultExcludes() {
-        return defaultExcludes;
-    }
-
-    /**
-     * <a href="https://ant.apache.org/manual/dirtasks.html#defaultexcludes" target="_blank">Ant style pattern</a>
-     * of files to disable default excludes from the archive.
-     *
-     * @return the defaultExcludes boolean
-     */
-    @DataBoundSetter
-    public void setDefaultExcludes(boolean defaultExcludes) {
-        this.defaultExcludes = defaultExcludes;
     }
 
 }
