+++ src/main/java/org/jenkinsci/plugins/pipeline/utility/steps/zip/ZipStep.java
@@ -34,6 +34,7 @@ import org.jenkinsci.plugins.workflow.steps.StepContext;
 import org.jenkinsci.plugins.workflow.steps.StepDescriptor;
 import org.jenkinsci.plugins.workflow.steps.StepExecution;
 import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.DataBoundSetter;
 
 import java.util.Set;
 
@@ -43,6 +44,29 @@ import java.util.Set;
  * @author Robert Sandell &lt;rsandell@cloudbees.com&gt;.
  */
 public class ZipStep extends AbstractFileCompressStep {
+
+    private boolean defaultExcludes = true;
+
+    /**
+     * <a href="https://ant.apache.org/manual/dirtasks.html#defaultexcludes" target="_blank">Ant style pattern</a>
+     * of files to disable default excludes from the archive.
+     *
+     * @return if defaultExcludes is enabled or not
+     */
+    public boolean isDefaultExcludes() {
+        return defaultExcludes;
+    }
+
+    /**
+     * <a href="https://ant.apache.org/manual/dirtasks.html#defaultexcludes" target="_blank">Ant style pattern</a>
+     * of files to disable default excludes from the archive.
+     *
+     * @param defaultExcludes if defaultExcludes is enabled or not
+     */
+    @DataBoundSetter
+    public void setDefaultExcludes(boolean defaultExcludes) {
+        this.defaultExcludes = defaultExcludes;
+    }
 
     @DataBoundConstructor
     public ZipStep(String zipFile) throws Descriptor.FormException {
