+++ src/main/resources/org/jenkinsci/plugins/pipeline/utility/steps/zip/ZipStep/help-defaultExclude.html
@@ -23,6 +23,6 @@
   -->
 
 <p>
-    <a href="https://ant.apache.org/manual/dirtasks.html#defaultexcludes" target="_blank">Ant style pattern</a>
-    of files to disable default excludes from the archive.
+    If the zip file should include defaultExcludes provided by ant java pattern.
+    By Default defaultExcludes param value will be enabled.
 </p>
